requirements.txt
opencv-contrib-python==4.0.0.21
pillow
h5py
scipy
numpy
tensorflow==2.4.1
scikit-learn
matplotlib
scikit-image
xlwt